requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
Flask-RESTful==0.3.9
flask-restx==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
MarkupSafe==2.1.1
pyrsistent==0.18.1
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
